+++ src/mode/visual_char_mode.rs
@@ -1706,6 +1706,7 @@ mod tests {
                 levels: Default::default(),
                 multiline: false,
                 continuation: None,
+                ..Default::default()
             })
             .unwrap(),
         )
@@ -1748,6 +1749,7 @@ mod tests {
             levels: Default::default(),
             multiline: false,
             continuation: None,
+            ..Default::default()
         };
         tab.display.format = Some(Arc::new(
             crate::parser::CustomParser::from_config(&cfg).unwrap(),